perf(sparkdown): resolve PageUp/PageDown locations on demand, drop pathLocations from the relay
pathLocations was the last heavy field riding along with every compile
notification — ~12k entries / ~470KB on a feature-length script — and it
existed solely to answer an occasional PageUp/PageDown keypress in the game
preview.

Move that lookup to the language server, which already owns the program:
new `sparkdown/offsetSourceLocation` request (same custom-request pattern as
sparkdown/fileReferences) returns the single {file, line} for a given
uri/line/offset. The client-side getOffsetSource/getClosestSourceIndex
helpers move to the server verbatim as getOffsetSourceLocation; PreviewGame
awaits the request instead of reading a mirrored program.

didCompile relay on the real ~8,300-line project: 600KB -> 126KB per edit.
Cumulative for this branch: 9,175KB -> 126KB (~73x).
